First try at sending raw opus data from Server to Clients and decoding them there

+++ src/main/java/kynake/audio/Sound.java
@@ -4,10 +4,10 @@ package kynake.audio;
 import net.minecraft.util.math.vector.Vector3d;
 
 public class Sound {
-  public final byte[] pcmSample;
+  public final short[] pcmSample;
   public final Vector3d sourceLocation;
 
-  public Sound(byte[] pcmSample, Vector3d sourceLocation) {
+  public Sound(short[] pcmSample, Vector3d sourceLocation) {
     this.pcmSample = pcmSample;
     this.sourceLocation = sourceLocation;
   }
